Fixed bootstrap issue with admin and reg user accounts
Admin and regular user account were not being correctly added during
bootstrap, which made testing the app difficult as I would have to
register a user every time I created the app.

Added fileOnError:true to the save command in each user account
creation, which led me to the issue of email not being added.  Email is
required now for every account.

Dummy email accounts added to the user creation, which fixed the issue.
Now able to log in using admin or user for different roles.